Add license for src files in rest frontend
