--- restful-order-service/frontend_server/src/sections/order/orders-search.js
@@ -1,24 +1,0 @@
-import MagnifyingGlassIcon from '@heroicons/react/24/solid/MagnifyingGlassIcon';
-import { Card, InputAdornment, OutlinedInput, SvgIcon } from '@mui/material';
-
-export const OrdersSearch = ({onChange, customer}) => (
-  <Card sx={{ p: 2 }}>
-    <OutlinedInput
-      fullWidth
-      placeholder="Filter by customer"
-      onChange={onChange}
-      value={customer}
-      startAdornment={(
-        <InputAdornment position="start">
-          <SvgIcon
-            color="action"
-            fontSize="small"
-          >
-            <MagnifyingGlassIcon />
-          </SvgIcon>
-        </InputAdornment>
-      )}
-      sx={{ maxWidth: 500 }}
-    />
-  </Card>
-);
